refactor(gateway): Q11.c — extract rest_v1/tasks/assignees.rs (GAR-635)
Extracts 3 assignee handlers + 2 DTOs (~333 LOC) from tasks/mod.rs into
tasks/assignees.rs (plan 0137, GAR-635 slice 3). mod.rs drops from 3180
to 2857 lines. Zero behaviour change; re-exports in mod.rs keep all
call-sites intact.
