+++ crates/garraia-gateway/src/rest_v1/openapi.rs
@@ -129,9 +129,9 @@ impl Modify for SecurityAddon {
         super::tasks::comments::create_task_comment,
         super::tasks::comments::list_task_comments,
         super::tasks::comments::delete_task_comment,
-        super::tasks::add_task_assignee,
-        super::tasks::list_task_assignees,
-        super::tasks::remove_task_assignee,
+        super::tasks::assignees::add_task_assignee,
+        super::tasks::assignees::list_task_assignees,
+        super::tasks::assignees::remove_task_assignee,
         super::tasks::create_task_label,
         super::tasks::list_task_labels,
         super::tasks::delete_task_label,
